requirements.txt
Flask==3.0.3
tensorflow-cpu
keras
numpy==1.26.4
Pillow==10.4.0
mysql-connector-python==9.1.0
Werkzeug==3.0.4
python-dotenv==1.0.1

